[routes]organise projects route folder
