+++ backend/src/routes/Projects.routes/checkIfUserIsOwner.ts
@@ -1,0 +1,29 @@
+import { Request, Response, NextFunction } from "express";
+import Project from "../../models/Project.models";
+import User from "../../models/User.models";
+
+export const checkIfUserIsOwner = async (
+  req: Request,
+  res: Response,
+  next: NextFunction
+) => {
+  try {
+    const project = await Project.findById(req.params.projectId);
+    if (!project) {
+      return res.status(404).json({ error: "Project not found" });
+    }
+    const { userId } = req.query;
+    const user = await User.findById(userId);
+    if (!user) {
+      return res.status(404).json({ error: "User not found" });
+    }
+
+    const isOwner = project.owner._id.toString() === user?._id!.toString();
+
+    res.json({
+      isOwner,
+    });
+  } catch (error) {
+    next(error);
+  }
+};
